Nest card network enum
